Began using `argparse` for \!commands parsing.
Got a working parser for the \!commands edit subcommand. Command names can now be any length and do not need to be passed to shlex. However, erroneous message saving and overwriting are possible due to limitations of the message body being positional. Consider requiring the message body be prefeced by '-m' so that regular `parse_args()` can be used.