Fix(parsing): Reworked how !cmd messages are parsed
`parse_cmd()` -> `parser.parse_args()`

I realized that, since the command name can only be one word and in an
absolute place at the beginning of the invocation (in fact it already
gets parsed outside of the parsing function), the name arg is no
longer necessary.
By adding a positional message arg, there is no longer a need for the
any part of the function. Note that the message nargs must be
argparse.REMAINDER to avoid parsing coincidental instances of the args
in the message body.
For some reason, the args with is_enabled and is_hidden dests, whose
actions set booleans, no longer defaulted to None. I suspect it's
because both args with set_true and set_false actions were moved to the
same parser. But I'm still not sure why. In any case, the defaults need
to be specified manually now.

Some other code was also cleaned up and error handling simplified.